[cleanup] Move the code for nodes in a dedicated folder like edges
