+++ packages/diagrams/frontend/sirius-components-diagrams-reactflow/src/renderer/node/ImageNode.types.ts
@@ -11,15 +11,9 @@
  *     Obeo - initial API and implementation
  *******************************************************************************/
 
-import { Label, NodeData } from './DiagramRenderer.types';
+import { NodeData } from '../DiagramRenderer.types';
 
-export interface ListNodeData extends NodeData {
-  style: React.CSSProperties;
-  listItems: ListItemData[];
-}
-
-export interface ListItemData {
-  id: string;
-  label: Label;
+export interface ImageNodeData extends NodeData {
+  imageURL: string;
   style: React.CSSProperties;
 }
